refactor(userspace): chisel utils
