+++ userspace/chisel/chisel_utils.h
@@ -1,0 +1,33 @@
+/*
+Copyright (C) 2013-2018 Draios Inc dba Sysdig.
+
+This file is part of sysdig.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+
+*/
+
+#pragma once
+
+///////////////////////////////////////////////////////////////////////////////
+// Initializer class.
+// An instance of this class is created when the library is loaded.
+// ONE-SHOT INIT-TIME OPERATIONS SHOULD BE DONE IN THE CONSTRUCTOR OF THIS
+// CLASS TO KEEP THEM UNDER A SINGLE PLACE.
+///////////////////////////////////////////////////////////////////////////////
+class chisel_initializer
+{
+public:
+	chisel_initializer();
+	~chisel_initializer();
+};
