+++ userspace/libsinsp/sinsp.cpp
@@ -59,10 +59,6 @@ limitations under the License.
 #include "tracer_emitter.h"
 #endif
 
-#ifdef HAS_CHISELS
-extern vector<chiseldir_info>* g_chisel_dirs;
-#endif
-
 void on_new_entry_from_proc(void* context, scap_t* handle, int64_t tid, scap_threadinfo* tinfo,
 							scap_fdinfo* fdinfo);
 
@@ -1829,32 +1825,6 @@ void sinsp::set_min_log_severity(sinsp_logger::severity sev)
 sinsp_evttables* sinsp::get_event_info_tables()
 {
 	return &g_infotables;
-}
-
-void sinsp::add_chisel_dir(string dirname, bool front_add)
-{
-#ifdef HAS_CHISELS
-	trim(dirname);
-
-	if(dirname[dirname.size() -1] != '/')
-	{
-		dirname += "/";
-	}
-
-	chiseldir_info ncdi;
-
-	ncdi.m_dir = std::move(dirname);
-	ncdi.m_need_to_resolve = false;
-
-	if(front_add)
-	{
-		g_chisel_dirs->insert(g_chisel_dirs->begin(), ncdi);
-	}
-	else
-	{
-		g_chisel_dirs->push_back(ncdi);
-	}
-#endif
 }
 
 void sinsp::set_buffer_format(sinsp_evt::param_fmt format)
